Hoàn thành thuật toán. đang bị lỗi hiển thị

+++ DTO/MT_DON_GIA.cs
@@ -10,7 +10,7 @@ namespace DTO
     {
         public int ID { get; set; }
         public string DIA_CHI { get; set; }
-        public int DON_GIA { get; set; }
+        public double DON_GIA { get; set; }
         public string GHI_CHU { get; set; }
     }
 }
